+++ vistock/plotly/prf4s.py
@@ -20,9 +20,11 @@ from plotly.subplots import make_subplots
 from .. import tw
 from .. import file_util
 from . import fig_util as futil
+from ..util import MarketColorStyle, decide_market_color_style
 
 
-def _plot(df, profile_field='Volume', period='12mo', interval='1d',
+def _plot(df, ticker, market_color_style, profile_field='Volume',
+          period='12mo', interval='1d',
           ma_nitems=(5, 10, 20, 50, 150), vma_nitems=50, total_bins=42,
           hides_nontrading=True):
     # Initialize empty plot with marginal subplots
@@ -40,10 +42,13 @@ def _plot(df, profile_field='Volume', period='12mo', interval='1d',
     )
 
     # Plot the candlestick chart
+    mc_style = decide_market_color_style(ticker, market_color_style)
+    mc_colors = futil.get_candlestick_colors(mc_style)
     candlestick = go.Candlestick(
         x=df.index,
         open=df['Open'], high=df['High'], low=df['Low'], close=df['Close'],
         name='OHLC',
+        **mc_colors
     )
     fig.add_trace(candlestick, row=1, col=1)
 
@@ -56,8 +61,9 @@ def _plot(df, profile_field='Volume', period='12mo', interval='1d',
         fig.add_trace(ma, row=1, col=1)
 
     # Add volume trace to 2nd row
-    colors = ['green' if c >= o
-            else 'red' for o, c in zip(df['Open'], df['Close'])]
+    cl = futil.get_volume_colors(mc_style)
+    colors = [cl['up'] if c >= o
+              else cl['down'] for o, c in zip(df['Open'], df['Close'])]
     volume = go.Bar(x=df.index, y=df['Volume'], name='Volume',
                     marker_color=colors)
     fig.add_trace(volume, row=2, col=1)
@@ -132,7 +138,8 @@ class Volume:
     """
     def plot(symbol='TSLA', period='12mo', interval='1d',
              ma_nitems=(5, 10, 20, 50, 150), vma_nitems=50, total_bins=42,
-             hides_nontrading=True, out_dir='out'):
+         hides_nontrading=True, market_color_style=MarketColorStyle.AUTO,
+         out_dir='out'):
         """Plot a price-by-volume, PBV  (also called volume profile) figure for
         a given stock. This figure shows the volume distribution across price
         levels for a stock.
@@ -181,16 +188,19 @@ class Volume:
             the number of bins to calculate comulative volume for bins.
         hides_nontrading: bool
             decide if hides non-trading time-periods.
+        market_color_style (MarketColorStyle): The market color style to use.
+            Default is MarketColorStyle.AUTO.
         out_dir: str
             the output directory for saving figure.
         """
         # Download stock data
-        symbol = tw.as_yfinance(symbol)
-        df = yf.Ticker(symbol).history(period=period, interval=interval)
+        ticker = tw.as_yfinance(symbol)
+        df = yf.Ticker(ticker).history(period=period, interval=interval)
 
         # Plot
-        fig = _plot(df, 'Volume', period, interval,
-                    ma_nitems, vma_nitems, total_bins, hides_nontrading)
+        fig = _plot(df, ticker, market_color_style, 'Volume',
+                    period, interval, ma_nitems, vma_nitems,
+                    total_bins, hides_nontrading)
         fig.update_layout(
             title=f'{symbol} {interval} '
                   f'({df.index.values[0]}~{df.index.values[-1]})',
@@ -214,7 +224,8 @@ class Turnover:
     '''
     def plot(symbol='TSLA', period='12mo', interval='1d',
              ma_nitems=(5, 10, 20, 50, 150), vma_nitems=50, total_bins=42,
-             hides_nontrading=True, out_dir='out'):
+         hides_nontrading=True, market_color_style=MarketColorStyle.AUTO,
+         out_dir='out'):
         """Plot a price-by-volume, PBV  (also called volume profile) figure for
         a given stock. This figure shows the volume distribution across price
         levels for a stock.
@@ -263,17 +274,20 @@ class Turnover:
             the number of bins to calculate comulative volume for bins.
         hides_nontrading: bool
             decide if hides non-trading time-periods.
+        market_color_style (MarketColorStyle): The market color style to use.
+            Default is MarketColorStyle.AUTO.
         out_dir: str
             the output directory for saving figure.
         """
         # Download stock data
-        symbol = tw.as_yfinance(symbol)
-        df = yf.Ticker(symbol).history(period=period, interval=interval)
+        ticker = tw.as_yfinance(symbol)
+        df = yf.Ticker(ticker).history(period=period, interval=interval)
         df['Turnover'] = df['Close'] * df['Volume']
 
         # Plot
-        fig = _plot(df, 'Turnover', period, interval,
-                    ma_nitems, vma_nitems, total_bins, hides_nontrading)
+        fig = _plot(df, ticker, market_color_style, 'Turnover',
+                    period, interval, ma_nitems, vma_nitems,
+                    total_bins, hides_nontrading)
         fig.update_layout(
             title=f'{symbol} {interval} '
                   f'({df.index.values[0]}~{df.index.values[-1]})',